Some more fixes on trigger time

+++ include/TTrigBit.hh
@@ -48,7 +48,7 @@ public:
     ap_int<32> ftrword;
     ap_int<32> ftrword2; // This is the 2nd trigger work front panel trigger input, after signal stretching;
     ap_int<48> ftrtime;  // This is a trigger time
-    long ftrig_time; // in 4ns steps
+    long long ftrig_time; // in 4ns steps
 
 private:
